requirements.txt
accelerate
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
annotated-types
anyio
attrs
azure-cognitiveservices-speech
azure-core
certifi
charset-normalizer
click
datasets
diffusers
dill
fastapi
filelock
frozenlist
fsspec
h11
hf-xet
httpcore
httpx
huggingface_hub
idna
importlib_metadata
Jinja2
MarkupSafe
mpmath
multidict
multiprocess
networkx
numpy
nvidia-cublas
nvidia-cuda-cupti
nvidia-cuda-nvrtc
nvidia-cuda-runtime
nvidia-cudnn-cu13
nvidia-cufft
nvidia-cufile
nvidia-curand
nvidia-cusolver
nvidia-cusparse
nvidia-cusparselt-cu13
nvidia-nccl-cu13
nvidia-nvjitlink
nvidia-nvshmem-cu13
nvidia-nvtx
packaging
pandas
peft
pillow
propcache
psutil
psycopg2-binary
pyarrow
pydantic
pydantic_core
python-dateutil
python-dotenv
pytz
PyYAML
regex
requests
safetensors
setuptools
shellingham
six
starlette
sympy
tokenizers
torch==2.9.1+cu130
torchvision==0.24.1+cu130
tqdm
transformers
triton
typing-inspection
typing_extensions
tzdata
urllib3
uvicorn
xxhash
yarl
zipp
